requirements.txt
scikit-learn
Pillow
matplotlib
numpy
